+++ src/frequenz/sdk/timeseries/__init__.py
@@ -5,17 +5,11 @@
 Handling of timeseries streams.
 
 A timeseries is a stream (normally an async iterator) of
-[samples][frequenz.sdk.timeseries.Sample].
-
-This module provides tools to operate on timeseries.
+[`Sample`][frequenz.sdk.timeseries.Sample]s.
 """
 
-from ._resampler import GroupResampler, Resampler, ResamplingFunction
-from ._sample import Sample
+from ._base_types import Sample
 
 __all__ = [
-    "GroupResampler",
-    "Resampler",
-    "ResamplingFunction",
     "Sample",
 ]