+++ src/frequenz/sdk/util/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
+
+"""General purpose tools."""